Base Attribute Buyer files modified slightly and added to repo.
So the BaseAttributeBuyer files have been added to the repository,
though slightly altered from their original form.  In particular, the
only script run in the html file is a call to BaseAttributeBuyer.js in
the head element, and then setting up all necessary events is done
through said .js file.  That, and the files now both use the jQuery
library.