Started working on the prototype 'back-end' for the Attribute Buyer module.
The current structure of the AttributeBuyer module has the view as its
own entity as of my last commit, while the controller and the model are
largely identified.  The Game.js file is the foundation for the
separation of the latter pair.  In detail: Game.js will define a
prototype which the controller will instantiate once on startup, then
ask for a new character.  Game builds the new character and returns it
to the controller which, in turn will take data from the Character
object and attach it to the appropriate places on the view.  As the user
generates input the controller will feed information into the character
which will occasionally update itself with some help from Game.  Most of
this is not implemented. My intention, for now, is to try to have Game
behave like a background universe.  This approach approximates both
itself (having a 'Game' or 'Rules' object full of various functions and
lists that character objects reference to build themselves) as well as
the database approach (where 'Game' is replaced by a database of data
which a character object can use to build itself).  My aim is to build
the infrastructure to be adaptable to both approaches.

GamePrototypeTests.htm is a page I'm using to test and debug Game.js.

README updated