Completed break between Model and Controller.  Added the capacity to create a new character from the View.
The BaseAttributeBuyer.js file now acts as a bridge between the html
file (the view(?)) and the Game.js file which contains all of the
star-traveller-specific guts.  The code executes as follows:

HTML elements load
BaseAttributeBuyer.js (Main) asks Game to create a pair of characters,
these characters are appended to a list.  Main then displays the
character 'names' on a table on the webpage.  The user can switch
between the two different characters from there.

After obtaining the new characters, Main "binds" the first of them to
the appropriate HTML elements.  Effectively displaying 'Character 0'.

As the user changes elements on the screen, the bound character's
appropriate attributes change, then the screen is updated.

If the user chooses a new character, their Attributes are "bound" and
displayed.

A button/code was added to create even more new characters.