Added infrastructure for adding character classes to the manager.  Corrected a typo in the dev_log.  Added my classes to index.html
Created the Class and Feat constructors.  These will mostly likely serve
as visual maps for building the individual class/feat objects.  I don't
expect to see these in the final project.  Added a json file with the
skeleton of the data necessary to define the "Medic" class.  Some
attributes were added to Game and Character to accommodate these
changes.  Status: Incomplete.

Corrected "June" to "July" in dev_log so that the proper log is
referenced.

Added references to Game.js and Character.js to the index.html file.