Refactored code generally.  Edited Game.js to reflect the removal of attribute buy points from the character object.
Multiple elements of code were refactored or compressed to reflect
design decisions made a few weeks ago.
In particular the Character object now does not keep track of the points
used to purchase its base attributes, rather this points value is
calculated.  This has prompted a change to the implementation of Game's
BuyAttributePointsForChar method.