Changed the Stat property of Character to make it dynamically computed.
The Stat Property now functions as follows:
Each stat is a property of Character.Stats.  Said property is itself an
object whose keys represent the sources contributing to the stat in
question, and whose values are the amounts by which said contributors
add to the stat in question.  The new method StatScore on the character
prototype asks Character.Stats[stat] for its list of contributors and
returns the sum of the associated values.